Split view and test files to group related endpoints
Splits views/auth into:
* views/auth - login/logout endpoints
* views/create_user - endpoints for completing user registration
* views/reset_password - endpoints for 'forgot password' request
  and setting the new password
