+++ app/main/__init__.py
@@ -3,4 +3,4 @@ from flask import Blueprint
 main = Blueprint('main', __name__)
 
 from . import errors
-from .views import auth, status
+from .views import auth, reset_password, create_user, status
