+++ app/main/views/auth.py
@@ -1,20 +1,12 @@
 # coding: utf-8
-from __future__ import unicode_literals
-from __future__ import absolute_import
-
-import six
-
 from flask_login import current_user
-from flask import abort, current_app, flash, redirect, render_template, request, url_for, get_flashed_messages, Markup
+from flask import current_app, flash, redirect, render_template, request, url_for, get_flashed_messages, Markup
 from flask_login import logout_user, login_user
 
-from dmapiclient import HTTPError
 from dmutils.user import User
-from dmutils.email import (decode_invitation_token, decode_password_reset_token, generate_token, send_email)
-from dmutils.email.exceptions import EmailError
 
 from .. import main
-from ..forms.auth_forms import LoginForm, EmailAddressForm, ChangePasswordForm, CreateUserForm
+from ..forms.auth_forms import LoginForm
 from ..helpers import hash_email
 from ..helpers.login_helpers import redirect_logged_in_user
 from ... import data_api_client
@@ -23,15 +15,6 @@ from ... import data_api_client
 NO_ACCOUNT_MESSAGE = Markup("""Make sure you've entered the right email address and password. Accounts
     are locked after 5 failed attempts. If you think your account has been locked, email
     <a href="mailto:enquiries@digitalmarketplace.service.gov.uk">enquiries@digitalmarketplace.service.gov.uk</a>.""")
-PASSWORD_UPDATED_MESSAGE = "You have successfully changed your password."
-PASSWORD_NOT_UPDATED_MESSAGE = "Could not update password due to an error."
-EMAIL_SENT_MESSAGE = Markup("""If the email address you've entered belongs to a Digital Marketplace account, we'll
-    send a link to reset the password.""")
-EXPIRED_PASSWORD_RESET_TOKEN_MESSAGE = Markup("""This password reset link has expired. Enter your email address and we’ll send you a
-    new one. Password reset links are only valid for 24 hours.""")
-INVALID_TOKEN_MESSAGE = Markup("""The link you used to create an account is not valid. Check you’ve entered the correct
-    link. If you still can’t create an account, email
-    <a href="mailto:enquiries@digitalmarketplace.service.gov.uk">enquiries@digitalmarketplace.service.gov.uk</a>""")
 
 
 @main.route('/login', methods=["GET"])
@@ -81,231 +64,3 @@ def process_login():
 def logout():
     logout_user()
     return redirect(url_for('.render_login'))
-
-
-@main.route('/reset-password', methods=["GET"])
-def request_password_reset():
-    return render_template("auth/request-password-reset.html",
-                           form=EmailAddressForm()), 200
-
-
-@main.route('/reset-password', methods=["POST"])
-def send_reset_password_email():
-    form = EmailAddressForm()
-    if form.validate_on_submit():
-        email_address = form.email_address.data
-        user_json = data_api_client.get_user(email_address=email_address)
-
-        if user_json is not None:
-
-            user = User.from_json(user_json)
-
-            token = generate_token(
-                {
-                    "user": user.id
-                },
-                current_app.config['SECRET_KEY'],
-                current_app.config['RESET_PASSWORD_SALT']
-            )
-
-            url = url_for('main.reset_password', token=token, _external=True)
-
-            email_body = render_template(
-                "emails/reset_password_email.html",
-                url=url,
-                locked=user.locked)
-
-            try:
-                send_email(
-                    user.email_address,
-                    email_body,
-                    current_app.config['DM_MANDRILL_API_KEY'],
-                    current_app.config['RESET_PASSWORD_EMAIL_SUBJECT'],
-                    current_app.config['RESET_PASSWORD_EMAIL_FROM'],
-                    current_app.config['RESET_PASSWORD_EMAIL_NAME'],
-                    ["password-resets"]
-                )
-            except EmailError as e:
-                current_app.logger.error(
-                    "Password reset email failed to send. "
-                    "error {error} email_hash {email_hash}",
-                    extra={'error': six.text_type(e),
-                           'email_hash': hash_email(user.email_address)})
-                abort(503, response="Failed to send password reset.")
-
-            current_app.logger.info(
-                "login.reset-email.sent: Sending password reset email for "
-                "supplier_id {supplier_id} email_hash {email_hash}",
-                extra={'supplier_id': user.supplier_id,
-                       'email_hash': hash_email(user.email_address)})
-        else:
-            current_app.logger.info(
-                "login.reset-email.invalid-email: "
-                "Password reset request for invalid supplier email {email_hash}",
-                extra={'email_hash': hash_email(email_address)})
-
-        flash(EMAIL_SENT_MESSAGE)
-        return redirect(url_for('.request_password_reset'))
-    else:
-        return render_template("auth/request-password-reset.html",
-                               form=form), 400
-
-
-@main.route('/reset-password/<token>', methods=["GET"])
-def reset_password(token):
-    decoded = decode_password_reset_token(token, data_api_client)
-    if 'error' in decoded:
-        flash(EXPIRED_PASSWORD_RESET_TOKEN_MESSAGE, 'error')
-        return redirect(url_for('.request_password_reset'))
-
-    email_address = decoded['email']
-
-    return render_template("auth/reset-password.html",
-                           email_address=email_address,
-                           form=ChangePasswordForm(),
-                           token=token), 200
-
-
-@main.route('/reset-password/<token>', methods=["POST"])
-def update_password(token):
-    form = ChangePasswordForm()
-    decoded = decode_password_reset_token(token, data_api_client)
-    if 'error' in decoded:
-        flash(EXPIRED_PASSWORD_RESET_TOKEN_MESSAGE, 'error')
-        return redirect(url_for('.request_password_reset'))
-
-    user_id = decoded["user"]
-    email_address = decoded["email"]
-    password = form.password.data
-
-    if form.validate_on_submit():
-        if data_api_client.update_user_password(user_id, password, email_address):
-            current_app.logger.info(
-                "User {user_id} successfully changed their password",
-                extra={'user_id': user_id})
-            flash(PASSWORD_UPDATED_MESSAGE)
-        else:
-            flash(PASSWORD_NOT_UPDATED_MESSAGE, 'error')
-        return redirect(url_for('.render_login'))
-    else:
-        return render_template("auth/reset-password.html",
-                               email_address=email_address,
-                               form=form,
-                               token=token), 400
-
-
-@main.route('/create/<string:encoded_token>', methods=["GET"])
-def create_user(encoded_token):
-    token = decode_invitation_token(encoded_token)
-
-    if token.get('error') == 'token_invalid':
-        current_app.logger.warning(
-            "createuser.token_invalid: {encoded_token}",
-            extra={'encoded_token': encoded_token}
-        )
-        abort(400, INVALID_TOKEN_MESSAGE)
-
-    role = token["role"]
-
-    if token.get('error') == 'token_expired':
-        current_app.logger.warning(
-            "createuser.token_expired: {encoded_token}",
-            extra={'encoded_token': encoded_token}
-        )
-        return render_template(
-            "auth/create-user-error.html",
-            error=None,
-            role=role,
-            token=None,
-            user=None), 400
-
-    form = CreateUserForm()
-
-    user_json = data_api_client.get_user(email_address=token["email_address"])
-
-    if not user_json:
-        return render_template(
-            "auth/create-user.html",
-            email_address=token['email_address'],
-            form=form,
-            role=role,
-            supplier_name=token.get('supplier_name'),
-            token=encoded_token), 200
-
-    user = User.from_json(user_json)
-    return render_template(
-        "auth/create-user-error.html",
-        error=None,
-        role=role,
-        token=token,
-        user=user), 400
-
-
-@main.route('/create/<string:encoded_token>', methods=["POST"])
-def submit_create_user(encoded_token):
-    token = decode_invitation_token(encoded_token)
-
-    if token.get('error') == 'token_invalid':
-        current_app.logger.warning(
-            "createuser.token_invalid: {encoded_token}",
-            extra={'encoded_token': encoded_token}
-        )
-        abort(400, INVALID_TOKEN_MESSAGE)
-
-    role = token["role"]
-
-    if token.get('error') == 'token_expired':
-        current_app.logger.warning(
-            "createuser.token_expired: {encoded_token}",
-            extra={'encoded_token': encoded_token}
-        )
-        return render_template(
-            "auth/create-user-error.html",
-            error=None,
-            role=role,
-            token=None,
-            user=None), 400
-
-    form = CreateUserForm()
-
-    if not form.validate_on_submit():
-        current_app.logger.warning(
-            "createuser.invalid: {form_errors}",
-            extra={'form_errors': ", ".join(form.errors)})
-        return render_template(
-            "auth/create-user.html",
-            email_address=token['email_address'],
-            form=form,
-            role=role,
-            supplier_name=token.get('supplier_name'),
-            token=encoded_token), 400
-
-    try:
-        user_data = {
-            'name': form.name.data,
-            'password': form.password.data,
-            'emailAddress': token['email_address'],
-            'role': role
-        }
-
-        if role == 'buyer':
-            user_data.update({'phoneNumber': form.phone_number.data})
-        elif role == 'supplier':
-            user_data.update({'supplierId': token['supplier_id']})
-
-        user_create_response = data_api_client.create_user(user_data)
-        user = User.from_json(user_create_response)
-        login_user(user)
-
-    except HTTPError as e:
-        if e.status_code == 409 or e.message == 'invalid_buyer_domain':
-            return render_template(
-                "auth/create-user-error.html",
-                error=e.message,
-                role=role,
-                token=None), 400
-        else:
-            raise
-
-    flash('account-created', 'flag')
-    return redirect_logged_in_user()
